inference/requirements.txt
evalutils==0.2.4
scikit-learn>=0.20.2
scipy>=1.2.1
SimpleITK==2.0.2
numpy 
olefile==0.46
packaging==21.3
Pillow==8.4.0
monai==0.8.0
nibabel
scikit-image
itk
pydicom==1.3.0
torchvision 
fill_voids